+++ src/Control/SubmitButton.php
@@ -4,12 +4,6 @@ namespace Efabrica\NeoForms\Control;
 
 class SubmitButton extends \Nette\Forms\Controls\SubmitButton
 {
-    public function __construct($caption = null)
-    {
-        parent::__construct($caption);
-        $this->setOption('type', 'submit');
-    }
-
     public function setIcon(?string $icon): self
     {
         $this->setOption('icon', $icon);